refactor home test and webapp engine locations
