+++ cmd/webapp.go
@@ -2,10 +2,8 @@ package cmd
 
 import (
 	"fmt"
-	"html/template"
 	"log"
 
-	"github.com/gin-gonic/gin"
 	"github.com/spf13/cobra"
 
 	"github.com/SUSE/console-for-sap/webapp"
@@ -35,18 +33,8 @@ func init() {
 	serveCmd.Flags().IntVarP(&port, "port", "p", 8080, "The port for the HTTP service to listen at")
 }
 
-func makeEngine() *gin.Engine {
-	templates := template.Must(template.New("").ParseFS(webapp.FS, "templates/*.tmpl"))
-
-	engine := gin.Default()
-	engine.SetHTMLTemplate(templates)
-	engine.GET("/", webapp.Home)
-
-	return engine
-}
-
 func serve(cmd *cobra.Command, args []string) {
-	engine := makeEngine()
+	engine := webapp.Engine()
 
 	listenAddress := fmt.Sprintf("%s:%d", host, port)
 	log.Fatal(engine.Run(listenAddress))
