--- webapp/embedded_fs.go
@@ -1,6 +1,0 @@
-package webapp
-
-import "embed"
-
-//go:embed templates
-var FS embed.FS
